Issue #640: move deconv padding calculation
